+++ database/model/model.go
@@ -23,18 +23,19 @@ type User struct {
 }
 
 type Client struct {
-	Id       uint            `json:"id" form:"id" gorm:"primaryKey;autoIncrement"`
-	Enable   bool            `json:"enable" form:"enable"`
-	Name     string          `json:"name" form:"name"`
-	Config   json.RawMessage `json:"config,omitempty" form:"config"`
-	Inbounds json.RawMessage `json:"inbounds" form:"inbounds"`
-	Links    json.RawMessage `json:"links,omitempty" form:"links"`
-	Volume   int64           `json:"volume" form:"volume"`
-	Expiry   int64           `json:"expiry" form:"expiry"`
-	Down     int64           `json:"down" form:"down"`
-	Up       int64           `json:"up" form:"up"`
-	Desc     string          `json:"desc" form:"desc"`
-	Group    string          `json:"group" form:"group"`
+	Id        uint            `json:"id" form:"id" gorm:"primaryKey;autoIncrement"`
+	Enable    bool            `json:"enable" form:"enable"`
+	Name      string          `json:"name" form:"name"`
+	SubSecret string          `json:"subSecret,omitempty" form:"subSecret" gorm:"index"`
+	Config    json.RawMessage `json:"config,omitempty" form:"config"`
+	Inbounds  json.RawMessage `json:"inbounds" form:"inbounds"`
+	Links     json.RawMessage `json:"links,omitempty" form:"links"`
+	Volume    int64           `json:"volume" form:"volume"`
+	Expiry    int64           `json:"expiry" form:"expiry"`
+	Down      int64           `json:"down" form:"down"`
+	Up        int64           `json:"up" form:"up"`
+	Desc      string          `json:"desc" form:"desc"`
+	Group     string          `json:"group" form:"group"`
 
 	// Delay start and periodic reset
 	DelayStart bool  `json:"delayStart" form:"delayStart" gorm:"default:false;not null"`