+++ sub/jsonService.go
@@ -97,8 +97,7 @@ func (j *JsonService) GetJson(subId string, format string) (*string, []string, e
 
 func (j *JsonService) getData(subId string) (*model.Client, []*model.Inbound, error) {
 	db := database.GetDB()
-	client := &model.Client{}
-	err := db.Model(model.Client{}).Where("enable = true and name = ?", subId).First(client).Error
+	client, err := (&SubService{}).getClientBySubId(subId)
 	if err != nil {
 		return nil, nil, err
 	}